requirements.txt
# MINIMAL REQUIREMENTS FOR DIGITAL OCEAN DEPLOYMENT  
# Essential dependencies only - optimized for memory usage

# Core Framework (Essential)
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database & ORM (Essential)
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
asyncpg==0.29.0

# Redis Cache (Essential)
redis==5.0.1

# Authentication & Security (Essential)
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
cryptography==41.0.7
python-multipart==0.0.6
bcrypt==4.1.2

# HTTP Client & WebSocket (Essential)
httpx==0.25.2
websockets==12.0

# Crypto Trading APIs (Essential)
ccxt==4.1.60
python-binance==1.0.19

# Basic Data Processing (Lightweight)
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3

# System Monitoring & Performance (Essential)
psutil==5.9.6

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Date/Time Utilities
python-dateutil==2.8.2
pytz==2023.3

# CORS & Security
python-multipart==0.0.6

# JSON handling
ujson==5.10.0

# Additional lightweight utilities (Essential for API functionality)
aiohttp==3.9.1
aioredis==2.0.1
alembic==1.12.1
passlib[bcrypt]==1.7.4

# Lightweight ML essentials (Required by core system files)
scipy==1.11.4

# Development/Testing
pytest==7.4.3
pytest-asyncio==0.21.1